remove mumlibs dependency on rust-mumble-protocol
