+++ mumlib/Cargo.toml
@@ -11,6 +11,5 @@ edition = "2018"
 colored = "2.0"
 fern = "0.5"
 log = "0.4"
-mumble-protocol = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 toml = "0.5"
